requirements.txt
discord.py
python-dotenv==0.10.5
